+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-04-26 19:08:30
-// Commit: test(ui): add unit tests for ui business logic
+// Updated at: 2026-04-29 17:16:24
+// Commit: feat(ui): implement responsive ui view component
 
 class uiComponentState {
-  final String commitRef = '723';
-  final String lastMessage = 'test(ui): add unit tests for ui business logic';
-  final DateTime timestamp = DateTime.parse('2026-04-26T19:08:30+05:00');
+  final String commitRef = '766';
+  final String lastMessage = 'feat(ui): implement responsive ui view component';
+  final DateTime timestamp = DateTime.parse('2026-04-29T17:16:24+05:00');
 }